# openapi.yaml
openapi: 3.0.3
info:
  title: User Management API Specification
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/license/mit/
servers:
  - url: http://localhost:5289
    description: Development Server
tags:
  - name: User
    description: User
components:
  schemas:
    Email:
      type: object
      required:
        - address
      properties:
        address:
          type: string
        description:
          type: string
    User:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        name:
          $ref: '#/components/schemas/UserName'
        contact:
          $ref: '#/components/schemas/UserContact'
    UserName:
      type: object
      properties:
        first:
          type: string
        middle:
          type: string
        last:
          type: string
        display:
          type: string
    UserContact:
      type: object
      properties:
        emailAddresses:
          type: array
          items:
            $ref: '#/components/schemas/Email'
    UserListGetResponse:
      type: object
      required:
        - totalCount
        - count
        - users
      properties:
        totalCount:
          type: integer
          format: int64
        count:
          type: integer
          format: int64
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
    UserGetResponse:
      allOf:
        - $ref: '#/components/schemas/User'
    UserCreateResponse:
      type: object
      properties:
        id:
          type: string
    UserUpdateRequest:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/UserName'
        contact:
          $ref: '#/components/schemas/UserContact'
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: API-ACCESS-KEY
    basicAuth:
      type: http
      scheme: basic
paths:
  /users:
    get:
      tags:
        - User
      summary: summary of get
      description: description
      operationId: getUserList
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListGetResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '503':
          description: Service Unavailable
    post:
      tags:
        - User
      summary: summary of post
      description: description
      operationId: createUser
      responses:
        '200':
            description: success
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/UserCreateResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '503':
          description: Service Unavailable
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - User
      summary: summary of get
      description: description
      operationId: getUser
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGetResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '503':
          description: Service Unavailable
    put:
      tags:
        - User
      summary: summary of put
      description: description
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '202':
          description: accepted
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '503':
          description: Service Unavailable
    delete:
      tags:
        - User
      summary: summary of delete
      description: description
      operationId: deleteUser
      responses:
        '202':
          description: accepted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '503':
          description: Service Unavailable